Refactor-Opt: Use an atomic boolean to indicate stealing.
Since StealTask is protected by a mutex. There can be
only one stealer at a time. A boolean suffices to indicate
if the stealer is active. Also we don't need to use
fetch_add to inc and compare_exchange to dec. We can
use store to toggle the flag between true/false.
We make the flag private and use helper method in
sexecute_unit to read it.

Signed-off-by: Lilith Oberhauser <lilith.oberhauser@huawei.com>